requirements.txt
distlib==0.3.8
Django==5.1.1
django-allauth==64.2.1
django-filter==24.3
django-oauth-toolkit==3.0.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
jwcrypto==1.5.6
oauthlib==3.2.2
platformdirs==4.3.3
pycparser==2.22
PyJWT==2.9.0
requests==2.32.3
sqlparse==0.5.1
urllib3==2.2.2
phonenumbers==8.13.45
pycountry==24.6.1
